application_controller.rbに記述していたレベル処理をconcerns/level_up_down.rbに移動 #190
